[libpas] Parallelize test runner https://bugs.webkit.org/show_bug.cgi?id=304285 rdar://164655522
Reviewed by Keith Miller.

With this patch, TestHarness.cpp will launch up to NCPUs/2 test
processes at once by default. This is friendlier than NCPUs for
contended cases like CI runners, while still being a big speedup
for desk usage.
This default can be overridden via the PasTestConcurrency envvar.

On my M3 Macbook Pro, this reduces the wall-clock time of `./test.sh`
from ~16m to a bit under ~5m, using `PasTestConcurrency=8`.

Canonical link: https://commits.webkit.org/305304@main

+++ Source/bmalloc/libpas/src/libpas/pas_mte.h
@@ -201,7 +201,7 @@ typedef enum pas_mte_tag_constraint pas_mte_tag_constraint;
 
 PAS_ALWAYS_INLINE pas_mte_tag_constraint pas_mte_exclude_tag(pas_mte_tag_constraint base, uint8_t tag_value_to_exclude)
 {
-    return (pas_mte_tag_constraint)(base & ~(1 << tag_value_to_exclude));
+    return (pas_mte_tag_constraint)((unsigned)base & ~(1u << tag_value_to_exclude));
 }
 
 PAS_ALWAYS_INLINE pas_mte_tag_constraint
